requirements.txt
pyTelegramBotAPI
python-telegram-bot
python-dotenv
fastapi 
uvicorn[standard]